# public/swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Collaborators API",
        "version": "1.0"
    },
    "paths": {
        "/api/collaborators": {
            "get": {
                "tags": [
                    "Collaborators"
                ],
                "summary": "Display a listing of collaborators",
                "operationId": "3bad7de0da48224fdc7f177841872903",
                "responses": {
                    "200": {
                        "description": "List of all collaborators",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Collaborator"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Collaborators"
                ],
                "summary": "Store a newly created collaborator",
                "operationId": "95299406c39c1533f68255ccedfe7d9b",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "user_id"
                                ],
                                "properties": {
                                    "user_id": {
                                        "description": "ID of the user",
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Collaborator created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Collaborator"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid input"
                    }
                }
            }
        },
...
        "/api/guests": {
            "get": {
                "tags": [
                    "Guests"
                ],
                "summary": "Display a listing of the resource",
                "operationId": "ba13faba9c30d75ce7386db1cee72e22",
                "responses": {
                    "200": {
                        "description": "List of guests",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Guest"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                }
            },
            "post": {
                "tags": [
                    "Guests"
                ],
                "summary": "Store a newly created resource in storage",
                "operationId": "abf1a99cd422a8b8b54c4a4ff63461f5",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/GuestRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Guest created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Guest"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid input"
                    }
                }
            }
        },
        "/api/guests/{id}": {
            "put": {
                "tags": [
                    "Guests"
                ],
                "summary": "Update the specified resource in storage",
                "operationId": "23fcd53cd995340d5018b111a6f722fb",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the guest",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/GuestRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Guest updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Guest"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid input"
                    },
                    "404": {
                        "description": "Guest not found"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Guests"
                ],
                "summary": "Remove the specified resource from storage",
                "operationId": "40fd2e7cb98f2852adbc8d7c4d5c8ee6",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the guest",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Guest deleted successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Guest deleted successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Guest not found"
                    }
                }
            }
        },
...
        "/api/posts": {
            "get": {
                "tags": [
                    "Posts"
                ],
                "summary": "Display a listing of the posts",
                "operationId": "e9eaf7df702a94f4443677bd1a36cfe6",
                "responses": {
                    "200": {
                        "description": "List of posts",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Post"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                }
            },
            "post": {
                "tags": [
                    "Posts"
                ],
                "summary": "Store a newly created post in storage",
                "operationId": "4a657e51a0066c64781851f2fd2651e2",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/PostRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Post created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Post"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid input"
                    }
                }
            }
        },
...
        "/api/tasks": {
            "get": {
                "tags": [
                    "Tasks"
                ],
                "summary": "Display a listing of tasks",
                "operationId": "471d938a3bfaeb6fbd981d7c3e2728b0",
                "responses": {
                    "200": {
                        "description": "List of tasks",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Task"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Tasks"
                ],
                "summary": "Store a newly created task",
                "operationId": "f803ebdb2300acdb7e93bbe8d0b9a2e8",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/TaskRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Task created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Task"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid input"
                    }
                }
            }
        },
...
        "/api/user": {
            "get": {
                "tags": [
                    "User"
                ],
                "summary": "Get the authenticated user's ID",
                "operationId": "6841f32423f371692de90b6e90854b3e",
                "responses": {
                    "200": {
                        "description": "Authenticated user's ID",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 1
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
...
            "Collaborator": {
                "title": "Collaborator",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "post_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "user_id": {
                        "type": "integer",
                        "example": 2
                    }
                },
                "type": "object"
            },
            "Guest": {
                "title": "Guest",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "John"
                    },
                    "surname": {
                        "type": "string",
                        "example": "Doe"
                    },
                    "phone": {
                        "type": "string",
                        "example": "+123456789"
                    },
                    "email": {
                        "type": "string",
                        "example": "guest@example.com"
                    },
                    "status": {
                        "type": "string",
                        "example": "invited"
                    },
                    "token": {
                        "type": "string",
                        "example": "abc123"
                    },
                    "user_id": {
                        "type": "integer",
                        "example": 2
                    }
                },
                "type": "object"
            },
            "Post": {
                "title": "Post",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "user_id": {
                        "type": "integer",
                        "example": 2
                    },
                    "title": {
                        "type": "string",
                        "example": "Wedding Event"
                    },
                    "description": {
                        "type": "string",
                        "example": "A beautiful wedding."
                    },
                    "wedding_date": {
                        "type": "string",
                        "format": "date",
                        "example": "2025-06-15"
                    },
                    "venue_name": {
                        "type": "string",
                        "example": "Grand Hall"
                    },
                    "venue_address": {
                        "type": "string",
                        "example": "123 Main St"
                    },
                    "latitude": {
                        "type": "number",
                        "format": "float",
                        "example": 40.7128
                    },
                    "longitude": {
                        "type": "number",
                        "format": "float",
                        "example": -74.006
                    }
                },
                "type": "object"
            },
            "Task": {
                "title": "Task",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "wedding_id": {
                        "type": "integer",
                        "example": 5
                    },
                    "title": {
                        "type": "string",
                        "example": "Book a venue"
                    },
                    "description": {
                        "type": "string",
                        "example": "Find and book the best venue"
                    },
                    "status": {
                        "type": "string",
                        "example": "pending"
                    },
                    "responsible_user_id": {
                        "type": "integer",
                        "example": 3
                    },
                    "deadline": {
                        "type": "string",
                        "format": "date",
                        "example": "2025-05-10"
                    }
                },
                "type": "object"
            }
        }
    },
...
}